+++ internal/config/keys/keys.go
@@ -21,7 +21,13 @@ const (
 	MachineID         = "machine_id"
 
 	// frequency of polling for actions.
-	PollingInterval = "poll_interval"
+	ActionPollingInterval = "action_poll_interval"
+
+	// frequency of tracking config checksums.
+	TrackerPollingInterval = "tracker_poll_interval"
+
+	// frequency of gathering agent status.
+	StatusPollingInterval = "status_poll_interval"
 
 	// AWS EC2 tags to be included in registration meta data.
 	AWSEC2Tags = "aws_ec2_tags"