+++ internal/platform/platform.go
@@ -3,12 +3,12 @@
 // license that can be found in the LICENSE file.
 //
 
-package agents
+package platform
 
 import "runtime"
 
 // getPlatform returns the OS and for darwin appends the architecture.
-func getPlatform() string {
+func Get() string {
 	platform := runtime.GOOS
 	if platform == "darwin" {
 		platform += "_" + runtime.GOARCH